Работа над ошибками 4.15. Добавлена модуляция 5.15
