+++ js/const.js
@@ -1,0 +1,27 @@
+export const PHOTOS_COUNT = 25;
+export const MAX_COMMENTS_COUNT = 3;
+export const MAX_RANGE_RANDOM_ID_COMMENT = 1000;
+export const MIN_LIKES_COUNT = 15;
+export const MAX_LIKES_COUNT = 200;
+export const MIN_AVATARS_COUNT = 1;
+export const MAX_AVATARS_COUNT = 6;
+
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export const NAMES = [
+  'Пётр',
+  'Алексей',
+  'Иван',
+  'Степан',
+  'Виктор',
+  'Александр',
+  'Константин',
+  'Всеволод',
+];
